refactor: 헤더 메뉴 리팩토링
