+++ frontend/src/components/header/header.const.js
@@ -1,0 +1,7 @@
+export const BANKS = [
+  { name: "KB국민은행", path: "/kb-bank" },
+  { name: "NH농협은행", path: "/nh-bank" },
+  { name: "신한은행", path: "/shinhan-bank" },
+  { name: "우리은행", path: "/woori-bank" },
+  { name: "하나은행", path: "/hana-bank" },
+];
