+++ frontend/src/components/index.js
@@ -1,1 +1,3 @@
 export * from "./header";
+export { default as Information } from "./Information";
+export { default as BankSection } from "./BankSection";
