--- frontend/src/pages/ProductPage/index.js
@@ -1,4 +1,0 @@
-export { CheckCard } from "./CheckCard";
-export { Deposit } from "./Deposit";
-export { Saving } from "./Saving";
-export { Loan } from "./Loan";
