+++ frontend/src/pages/index.js
@@ -1,12 +1,11 @@
+export { Admin } from "./Admin";
 export { default as BankPage } from "./BankPage";
 export { default as Container } from "./Container";
-export { Login } from "./Login";
-export { SignUp } from "./SignUp";
-export { Main } from "./Main";
 export { Detailed } from "./Detailed";
-export { ReviewForm } from "./ReviewForm";
+export { Login } from "./Login";
+export { Main } from "./Main";
 export { MyPage } from "./MyPage";
-export { Admin } from "./Admin";
+export { Product } from "./Product";
 export { ProductManagement } from "./ProductManagement";
-
-export * from "./ProductPage";
+export { ReviewForm } from "./ReviewForm";
+export { SignUp } from "./SignUp";
